+++ changelog.md
@@ -4,6 +4,7 @@ v4.0.7
 * Gave Phantom Membranes some moon affinity
 * Fixed tooltip Z-order issue with Runecarving Table screen
 * Fixed tooltip Z-order issue with Research Table screen
+* Changed research projects to use the Forge milk tag instead of just the milk bucket
 
 v4.0.6
 ------